FaceAttend_Web/requirements.txt
# FaceAttend Pro — Web App  |  pip install -r requirements.txt

# Web framework
flask>=3.0.0

# MySQL driver
mysql-connector-python>=8.3.0

# Computer Vision (must be contrib for cv2.face LBPH)
opencv-contrib-python>=4.8.0

# Image processing
Pillow>=10.0.0

# Numerical
numpy>=1.24.0
